move get_isle_compilations to meta::isle
